# Improve the performance of `BaseToQuoteQuantums` & `BigRatMulPpm`