### Changelist
- These are two of the most used function by the CPU at the moment.
- `BaseToQuoteQuantums` optimized by about 80% and `BigRatMulPpm` optimized by about 20%
- In a future PR, `multiplyByPrice` and `FundingRateToIndex` can be removed like in https://github.com/dydxprotocol/v4-chain/pull/1528, but am keeping this PR small for now to ensure it can be backported if needed

### Test Plan
- New benchmark test for `BaseToQuoteQuantums`
- Unit tests already exist for `BaseToQuoteQuantums` including rounding for negative numbers

### Benchmark Results

Before:
```
BenchmarkBigRatMulPpm-4          	16975401	       382.0 ns/op	     208 B/op	       9 allocs/op
BenchmarkBaseToQuoteQuantums-4   	 3997749	      1518 ns/op	    1056 B/op	      36 allocs/op
```

After:
```
BenchmarkBigRatMulPpm-4          	22503729	       289.2 ns/op	     192 B/op	       7 allocs/op
BenchmarkBaseToQuoteQuantums-4   	19702170	       266.7 ns/op	     184 B/op	       7 allocs/op

BenchmarkBigRatMulPpm-4: 24.3% improvement
BenchmarkBaseToQuoteQuantums-4: 82.4% improvement
```